[BugFix] Fix data parallel initialization and device management in NPUWorker and DPEngineCoreProc Co-authored-by: rjg-lyh <rjg-lyh@users.noreply.github.com>
